third_party/WebKit/LayoutTests/fast/animation/request-animation-frame-timestamps-advance-expected.txt

Tests the timestamps provided to requestAnimationFrame callbacks advance

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS firstTimestamp is defined.
PASS secondTimestamp is defined.
PASS secondTimestamp > firstTimestamp is true
PASS successfullyParsed is true

TEST COMPLETE
